Remove //gpu and GLES* headers from third_party/khronos
This gets rid of the rest of the files in //gpu and the GLES{2,3}
headers from //third_party/khronos. The GL headers used for GLES{2,3}
come from the platform being targetted, that is
//mojo/public/c/gpu/GLES2 for Mojo,
//third_party/android_tools/ndk/platforms/android-XX/usr/include/GLES2
for Android and the OpenGL framework for Mac/iOS.

The //third_party/khronos/EGL headers are still being used by some
utility code in //ui/gl. This code is only really useful on Android, as
Sky doesn't use EGL on iOS/Mac/Mojo, but it's compiled in on all
platforms currently and relies on some EGL defines that are not set in
the EGL headers we currently use. These could probably be removed
without much trouble.

The //third_party/khronos/GLES2 headers are still used by some of
//ui/gl on iOS. These should be using the iOS OpenGL framework headers,
and will in the future.

Removes the unneeded KHRONOS_APICALL override from KHR/khrplatform.h

Rolls to a newer Mojo SDK that doesn't expect //third_party/khronos